+++ src/index.ts
@@ -1,2 +1,3 @@
 export { createForme } from './forme/creator';
+export { createFormeFactory } from './forme/factory';
 export { initStore } from './store/creator';