+++ src/typing/internal.ts
@@ -21,6 +21,7 @@ export type State = any;
 export type Payload = any;
 export type RootState = any;
 export type Dispatch = any;
+export type CustomForme = any;
 export type Data = any;
 
 export enum ReFormeBuilderFields {
@@ -196,6 +197,12 @@ export interface ReFormeBuilder<
 }
 
 export type ReFormeBuilders = Record<string, ReFormeBuilder>;
+
+// Forme factory types
+
+export type Plugin = (customForme: CustomForme) => CustomForme;
+
+export type ReFactory = (customForme: CustomForme) => ReFormeBuilder;
 
 // Store types